Fixed most simple msvc warnings

+++ src/audio/adxaudiostream.cpp
@@ -76,13 +76,13 @@ bool AdxAudioStream::DecodeBuffer() {
 
     for (int i = 0; i < SamplesPerBuffer; i++) {
       /* this byte contains nibbles for two samples */
-      int sample_byte = input[i / 2];
+      uint8_t sample_byte = input[i / 2];
 
-      output[i * ChannelCount + c] =
-          clamp16((i & 1 ? get_low_nibble_signed(sample_byte)
-                         : get_high_nibble_signed(sample_byte)) *
-                      scale +
-                  (Coef1 * Hist1[c] >> 12) + (Coef2 * Hist2[c] >> 12));
+      output[i * ChannelCount + c] = (int16_t)(clamp16(
+          (i & 1 ? get_low_nibble_signed(sample_byte)
+                 : get_high_nibble_signed(sample_byte)) *
+              scale +
+          (Coef1 * Hist1[c] >> 12) + (Coef2 * Hist2[c] >> 12)));
 
       Hist2[c] = Hist1[c];
       Hist1[c] = output[i * ChannelCount + c];
