+++ src/data/savesystem.h
@@ -112,7 +112,7 @@ class SaveSystemBase {
   virtual void GetEVStatus(int evId, int* totalVariations,
                            int* viewedVariations) = 0;
   virtual void SetEVStatus(int id) = 0;
-  virtual bool GetEVVariationIsUnlocked(int evId, int variationIdx) = 0;
+  virtual bool GetEVVariationIsUnlocked(size_t evId, size_t variationIdx) = 0;
   virtual bool GetBgmFlag(int id) = 0;
   virtual void SetBgmFlag(int id, bool flag) = 0;
   virtual void SetCheckpointId(int id) = 0;
@@ -164,7 +164,7 @@ void GetReadMessagesCount(int* totalMessageCount, int* readMessageCount);
 void GetViewedEVsCount(int* totalEVCount, int* viewedEVCount);
 void GetEVStatus(int evId, int* totalVariations, int* viewedVariations);
 void SetEVStatus(int id);
-bool GetEVVariationIsUnlocked(int evId, int variationIdx);
+bool GetEVVariationIsUnlocked(size_t evId, size_t variationIdx);
 bool GetBgmFlag(int id);
 void SetBgmFlag(int id, bool setFlag);
 void SetCheckpointId(int id);
