+++ src/games/mo6tw/savesystem.h
@@ -45,7 +45,7 @@ class SaveSystem : public SaveSystemBase {
   void GetEVStatus(int evId, int* totalVariations,
                    int* viewedVariations) override;
   void SetEVStatus(int evId) override;
-  bool GetEVVariationIsUnlocked(int evId, int variationIdx) override;
+  bool GetEVVariationIsUnlocked(size_t evId, size_t variationIdx) override;
   bool GetBgmFlag(int id) override;
   void SetBgmFlag(int id, bool flag) override;
   void SetCheckpointId(int id) override {}
